refactor: core module
