+++ src/lib/index.ts
@@ -1,3 +1,3 @@
 export { ViewerComponent } from './src/viewer/viewer.component';
-export { MimeViewerIntl } from './src/viewer/viewer-intl';
+export { MimeViewerIntl } from './src/core/viewer-intl';
 export { MimeModule } from './src/module';
