+++ src/lib/src/contents-dialog/contents-dialog.component.html
@@ -11,7 +11,7 @@
     <div *ngSwitchDefault>
       <md-toolbar>
         <div fxLayout="row" fxLayoutAlign="space-between center" fxFlex>
-          <div class="heading">Contents</div>
+          <div class="heading heading-desktop">Contents</div>
           <button md-icon-button [attr.aria-label]="intl.close" [mdTooltip]="intl.close" [md-dialog-close]="true"><md-icon>close</md-icon></button>
         </div>
       </md-toolbar>
